requirements.txt
asgiref==3.8.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
google-ads==24.0.0
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
googleads==41.0.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
idna==3.7
isodate==0.6.1
lxml==5.2.1
numpy==1.26.4
oauthlib==3.2.2
platformdirs==4.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyodbc==5.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
xmltodict==0.13.0
zeep==4.2.1
